[RBP-026] Music Prefab 적용 미리 만들어놓은 Music Prefab를 적용해서 관리에 적합하도록 만들 예정. string 값에 따라 해당 string 값에 맞는 곡 / 노트 패턴으로 바뀌도록 만듬
