+++ RhythmBeatPlay/Assets/Scenes/GameScene.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 58c7af010007ecc4789ec5970e618594
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
