+++ RhythmBeatPlay/Assets/Script/Common/MusicDataHolder.cs
@@ -31,18 +31,6 @@ public class MusicDataHolder : MonoBehaviour
         DontDestroyOnLoad(this.transform);
     }
 
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
-
     public void ChangeMusicData(MusicData new_data)
     {
         data = new_data;
