+++ RhythmBeatPlay/Assets/Script/GameScene/note.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 07a17079289712d468b2b6b1d73bb162
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
